Removed violin unit test not needed and violin.bins not used anymore
